Split commands, refactor entry point.
