+++ crates/rb-cli/src/commands/mod.rs
@@ -1,17 +1,21 @@
 pub mod config;
 pub mod environment;
 pub mod exec;
+pub mod help;
 pub mod init;
 pub mod run;
 pub mod runtime;
 pub mod shell_integration;
 pub mod sync;
+pub mod version;
 
 pub use config::config_command;
 pub use environment::environment_command;
 pub use exec::exec_command;
+pub use help::help_command;
 pub use init::init_command;
 pub use run::run_command;
 pub use runtime::runtime_command;
 pub use shell_integration::shell_integration_command;
 pub use sync::sync_command;
+pub use version::version_command;
